+++ examples/features-check/src/core/services/theme/ThemeService.ts
@@ -39,7 +39,7 @@ export class ThemeService extends AbstractService {
     );
 
     // [*] Pass safe lifecycle checks - can emit from activation.
-    this.emitSignal({ type: `activated/${this.constructor.name}` });
+    this.emitSignal(`activated/${this.constructor.name}`);
   }
 
   @OnDeactivation()
@@ -47,16 +47,13 @@ export class ThemeService extends AbstractService {
     console.info(`[${this.constructor.name}] Deactivating`);
 
     // [*] Pass safe lifecycle checks - can emit from deactivation.
-    this.emitSignal({ type: `deactivating/${this.constructor.name}` });
+    this.emitSignal(`deactivating/${this.constructor.name}`);
   }
 
   @Action()
   public toggle(): void {
     this.theme = this.theme === "light" ? "dark" : "light";
 
-    this.emitSignal({
-      type: EGlobalSignal.THEME_TOGGLED,
-      payload: { theme: this.theme },
-    });
+    this.emitSignal(EGlobalSignal.THEME_TOGGLED, { theme: this.theme });
   }
 }