+++ src/wirestate/core/signals/use-signal-emitter.ts
@@ -7,24 +7,25 @@ import { prefix } from "@/macroses/prefix.macro";
 import { useIocContext } from "@/wirestate/core/provision/use-ioc-context";
 import { SIGNAL_BUS_TOKEN } from "@/wirestate/core/registry";
 import { SignalBus } from "@/wirestate/core/signals/signal-bus";
-import type { ISignal, TSignalEmitter } from "@/wirestate/types/signals";
+import type { TSignalEmitter, TSignalType } from "@/wirestate/types/signals";
 
 /**
  * Returns a stable function to emit signals.
  *
  * @returns signal emitter
  */
-export function useSignalEmitter(): TSignalEmitter {
+export function useSignalEmitter<P = unknown, T extends TSignalType = TSignalType>(): TSignalEmitter<P, T> {
   const container: Container = useIocContext().container;
 
   return useCallback(
-    (signal: ISignal) => {
+    <P, T extends TSignalType>(type: T, payload?: P, from?: unknown) => {
       dbg.info(prefix(__filename), "Emit signal:", {
-        type: signal?.type,
-        signal,
+        type,
+        payload,
+        from,
       });
 
-      container.get<SignalBus>(SIGNAL_BUS_TOKEN).emit(signal);
+      container.get<SignalBus>(SIGNAL_BUS_TOKEN).emit({ type, payload, from });
     },
     [container]
   );