+++ src/wirestate/types/signals.ts
@@ -1,3 +1,5 @@
+import { Optional } from "@/wirestate/types/general";
+
 /**
  * Signal identifier. Use symbols for private signals.
  */
@@ -6,9 +8,10 @@ export type TSignalType = string | symbol;
 /**
  * Signal object.
  */
-export interface ISignal<P = unknown, T extends TSignalType = TSignalType> {
+export interface ISignal<P = unknown, T extends TSignalType = TSignalType, F = unknown> {
   readonly type: T;
   readonly payload?: P;
+  readonly from?: F;
 }
 
 /**
@@ -27,7 +30,7 @@ export type TSignalUnsubscribe = () => void;
  * @internal
  */
 export interface ISignalDispatchEntry {
-  readonly types: ReadonlyArray<ISignal["type"]> | null;
+  readonly types: Optional<ReadonlyArray<ISignal["type"]>>;
   readonly handler: TSignalHandler;
 }
 
@@ -38,10 +41,14 @@ export interface ISignalDispatchEntry {
  */
 export interface ISignalHandlerMetadata {
   readonly methodName: string | symbol;
-  readonly types: ReadonlyArray<TSignalType> | null;
+  readonly types: Optional<ReadonlyArray<TSignalType>>;
 }
 
 /**
  * Signal emitter signature.
  */
-export type TSignalEmitter = (signal: ISignal) => void;
+export type TSignalEmitter<P = unknown, T extends TSignalType = TSignalType, F = unknown> = (
+  type: T,
+  payload?: P,
+  from?: F
+) => void;