+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `initialState` -> `seed`
 - `type` -> `bindingType`
 - `scopeType` -> `bindingScopeType`
+- Emit signal with two params instead of manually composed object, signal `from`
 
 ## 0.3.0